fix(export): accept float border_radius on pptx autoshape (#28)
PR #27's discriminated-union narrowed the old-deck pptx failure to
6 focused errors: PptxAutoShapeBoxModel.border_radius expects int,
receives floats like 10.5 or 274.57 from the Next.js converter.

The converter produces raw CSS pixel values from
element.getBoundingClientRect() — fractional by nature for
sub-pixel positioning. Downstream
apply_border_radius_to_shape passes this to python-pptx's Pt(),
which accepts float without issue (it just multiplies by 12700 to
get EMU), so the downstream consumer is already float-tolerant.

Widening the field to Optional[float] is the minimal fix and the
natural type for a CSS-derived quantity.

+++ servers/fastapi/models/pptx_models.py
@@ -128,7 +128,10 @@ class PptxAutoShapeBoxModel(PptxShapeModel):
     shadow: Optional[PptxShadowModel] = None
     position: PptxPositionModel
     text_wrap: bool = True
-    border_radius: Optional[int] = None
+    # Upstream converter emits raw CSS px values (e.g. 10.5, 274.57).
+    # Downstream apply_border_radius_to_shape passes this to
+    # python-pptx's Pt(), which accepts float without issue.
+    border_radius: Optional[float] = None
     paragraphs: Optional[List[PptxParagraphModel]] = None
